[JVM] TVMContext, TVMArray attributes

+++ jvm/core/src/main/scala/ml/dmlc/tvm/Shape.scala
@@ -3,25 +3,25 @@ package ml.dmlc.tvm
 /**
  * Shape of [[NDArray]] or other data
  */
-class Shape(dims: Traversable[Int]) extends Serializable {
+class Shape(dims: Traversable[Long]) extends Serializable {
   private val shape = dims.toVector
 
-  def this(dims: Int*) = {
+  def this(dims: Long*) = {
     this(dims.toVector)
   }
 
-  def apply(dim: Int): Int = shape(dim)
+  def apply(dim: Int): Long = shape(dim)
   def size: Int = shape.size
   def length: Int = shape.length
   def drop(dim: Int): Shape = new Shape(shape.drop(dim))
   def slice(from: Int, end: Int): Shape = new Shape(shape.slice(from, end))
-  def product: Int = shape.product
-  def head: Int = shape.head
+  def product: Long = shape.product
+  def head: Long = shape.head
 
   def ++(other: Shape): Shape = new Shape(shape ++ other.shape)
 
-  def toArray: Array[Int] = shape.toArray
-  def toVector: Vector[Int] = shape
+  def toArray: Array[Long] = shape.toArray
+  def toVector: Vector[Long] = shape
 
   override def toString(): String = s"(${shape.mkString(",")})"
 
@@ -37,7 +37,7 @@ class Shape(dims: Traversable[Int]) extends Serializable {
 }
 
 object Shape {
-  def apply(dims: Int *): Shape = new Shape(dims: _*)
-  def apply(dims: Traversable[Int]): Shape = new Shape(dims)
+  def apply(dims: Long *): Shape = new Shape(dims: _*)
+  def apply(dims: Traversable[Long]): Shape = new Shape(dims)
 }
